Particle date aliasing
Alias dates depending whether they are start or end points of a
date interval. For example, for end point dates, the following
are equivalent:

2015, 2015-12, 2015-12-31

It is important we alias these dates to their simplest form before
generating particle date solr queries. If we don't we can't match,
for example, a publication date specified as "2015-12" with the
within the date range ending with end date "2015-12-31".