Change test data after narrowing of available feed files
I've put some effrot into reducing the number of feed files that need to be processed
in order to have a passing test, this has resulted in some minor data difference that need to
be reflected in the test data. I've also changed the journal/{issn}/works endpoint test to pull
back all rows when asserting.