First attempt to implement CrossRef/rest-api-doc#327
This is an initial attempt to support the peer review content type. It is fully functioning
but the API currently exposes additional data that may not be required for peer reviews. This data
will be removed in a later commit.

No querying/filtering of the data has been tested/implemented yet.