Update readme with running tests info
As part of this commit I've also made the test fixture throw an exception if the
expected number of feed files is not present in the source directory `dev-resources/feeds/source`. We can
remove this once we commit the feed files to source control. I've done this because throwing an exception
and getting immediate feedback is better than waiting for Solr to finish indexing all of the expected files, which
will never happen.