Add tests around inconsistent work counts
Issues CrossRef/rest-api-doc/304 and CrossRef/rest-api-doc/324 report consistency
problems with works counts across a number of endpoints. This commit adds tests to verify
data returned from those endpoints and specifically to make sure work counts across endpoints
are consistent.

As part of this commit I also made a chage to make sure the `:items` collection asserted against during
testing was always in a deterministic state. This is why there are some changes to `dev-resources` that might not
be expected.

I also needed to mock out `get-member-list` and `get-prefix-info` to return member and prefix info of a known state.